[3787] Move attachment methods
