+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/factory/FileAttachmentFactory.kt
@@ -19,12 +19,11 @@ package io.getstream.chat.android.compose.ui.attachments.factory
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.runtime.Composable
-import com.getstream.sdk.chat.model.ModelType
-import io.getstream.chat.android.client.extensions.uploadId
 import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
 import io.getstream.chat.android.compose.ui.attachments.content.FileAttachmentContent
 import io.getstream.chat.android.compose.ui.attachments.content.FileAttachmentPreviewContent
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
+import io.getstream.chat.android.uiutils.extensions.isFile
 
 /**
  * An [AttachmentFactory] that validates attachments as files and uses [FileAttachmentContent] to
@@ -33,13 +32,7 @@ import io.getstream.chat.android.compose.ui.theme.ChatTheme
 @Suppress("FunctionName")
 public fun FileAttachmentFactory(): AttachmentFactory = AttachmentFactory(
     canHandle = { attachments ->
-        attachments.any {
-            it.uploadId != null ||
-                it.upload != null ||
-                it.type == ModelType.attach_file ||
-                it.type == ModelType.attach_video ||
-                it.type == ModelType.attach_audio
-        }
+        attachments.any { it.isFile() }
     },
     previewContent = @Composable { modifier, attachments, onAttachmentRemoved ->
         FileAttachmentPreviewContent(
