+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/factory/QuotedAttachmentFactory.kt
@@ -20,9 +20,9 @@ import androidx.compose.runtime.Composable
 import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
 import io.getstream.chat.android.compose.ui.attachments.content.FileAttachmentQuotedContent
 import io.getstream.chat.android.compose.ui.attachments.content.ImageAttachmentQuotedContent
-import io.getstream.chat.android.compose.ui.util.hasLink
-import io.getstream.chat.android.compose.ui.util.isFile
 import io.getstream.chat.android.compose.ui.util.isMedia
+import io.getstream.chat.android.uiutils.extensions.hasLink
+import io.getstream.chat.android.uiutils.extensions.isFile
 
 /**
  * An [AttachmentFactory] that validates attachments as files and uses [ImageAttachmentQuotedContent] in case the
