+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/AttachmentsUtils.kt
@@ -17,7 +17,6 @@
 package io.getstream.chat.android.compose.ui.util
 
 import com.getstream.sdk.chat.model.ModelType
-import io.getstream.chat.android.client.extensions.uploadId
 import io.getstream.chat.android.client.models.Attachment
 
 /**
@@ -35,28 +34,3 @@ private val MEDIA_ATTACHMENT_TYPES: Collection<String> = listOf(ModelType.attach
  * @return If the [Attachment] is media content or not.
  */
 internal fun Attachment.isMedia(): Boolean = type in MEDIA_ATTACHMENT_TYPES
-
-/**
- * @return If the [Attachment] is a link attachment or not.
- */
-internal fun Attachment.hasLink(): Boolean = titleLink != null || ogUrl != null
-
-/**
- * @return If the [Attachment] is a file or not.
- */
-internal fun Attachment.isFile(): Boolean {
-    return uploadId != null ||
-        upload != null ||
-        type == ModelType.attach_file ||
-        type == ModelType.attach_video ||
-        type == ModelType.attach_audio
-}
-
-/**
- * @return If the attachment is currently being uploaded to the server.
- */
-internal fun Attachment.isUploading(): Boolean {
-    return (uploadState is Attachment.UploadState.InProgress || uploadState is Attachment.UploadState.Idle) &&
-        upload != null &&
-        uploadId != null
-}
