+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/options/message/MessageOptionItemsFactory.kt
@@ -33,8 +33,8 @@ import io.getstream.chat.android.common.state.Resend
 import io.getstream.chat.android.common.state.ThreadReply
 import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.common.extensions.internal.getDrawableCompat
-import io.getstream.chat.android.ui.common.extensions.internal.hasLink
 import io.getstream.chat.android.ui.message.list.MessageListViewStyle
+import io.getstream.chat.android.uiutils.extensions.hasLink
 
 /**
  * An interface that allows the creation of message option items.
